Enabled the accordion to be properly editable via the admin panel
I added some code that removes the "slide" class from the sliding content classes if the page is "editable" and then also set the display to block mode to force it to show.

End result is that the slides are expanded in the admin panel and therefore properly ediable. As a result of this, I also removed the "Show more/Less" button as it was not aligned properly to start with, and could be seen as unnecessary now.