Added CV upload form and download in the email that is sent to the client
Also fixed a bug in the download controller for downloading brochures if the client uses a folder for their assets. This issue won't affect CV uploads as the folder for this will not change, but for the assets they will.

Fixed this passing the asset id to the controller instead of the filename, and using the asset class to get the path and name of the provided asset ID

YouTrack: #execcare-276 work 4h